Add support for alert relabeling
All other forms of relabeling (targets, metrics, remote write) are
already supported, but alert relabeling wasn't possible yet because it's
not part of the Prometheus Alertmanager configuration.

The most common use case for alert relabeling is dropping the "replica"
label of an HA Prometheus pair, so I chose that as an example.